refactor(src): Lotto 클래스 파일 위치 변경
