+++ src/main/kotlin/lotto/model/LottoNumberGenerator.kt
@@ -1,7 +1,5 @@
 package lotto.model
 
-import lotto.Lotto
-
 interface LottoNumberGenerator {
     fun generateNumbers(): Lotto
 }
